Complete renaming `Document` to `Record`
